refactor(generics): move SMPRequest to smp.__init__; drop smp.generics
The library needs only the Request/Response protocol; the TypeIs narrowing
helpers move into the static typing tests where the consumer contract is
asserted.
